requirements.txt
selenium
requests
python-dotenv
torch
transformers